Fixes #262
Add CSRF token support
